Correção nos métodos de comunicação com o banco de dados e nomenclatura das tabelas

+++ API/Entities/TB_Exercicios.cs
@@ -21,7 +21,7 @@ namespace API.Entities
         public int Tempo { get; set; }
         
         [Required]
-        public int UsuarioId { get; set; }
+        public long UsuarioId { get; set; }
         
         [ForeignKey("UsuarioId")]
         public virtual TB_USUARIO Usuario { get; set; }
